Stage two seems ready!
This is in need of refactoring, but functional (hopefully).

Looking at it myself, the results seem about right, so now I just need to implement stage 3, meaning: going over the file of "ad frames nearest to the long video's frames" and teach the computer how to interpret that. Woop woop.